Moved editor scripts.
