+++ Assets/Editor/FoodRoast.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 9ba492ab8f1f7ca4bae42bb22ac2c0db
+folderAsset: yes
+timeCreated: 1546130048
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
